pyproject.toml
[tool.poetry]
name = "playground"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
cocotb = "^1.9.1"
cocotbext-wishbone = "^0.2.2"
pytest = "^8.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
